+++ uart_hello/src/main.rs
@@ -4,7 +4,7 @@
 #![no_main]
 
 use hisi_panic_handler as _;
-use hisi_riscv_hal::{
+use hisi_hal::{
     peripherals::Peripherals,
     uart::{Config, Uart, UartClock},
 };